BUG: find_objects bool support
* Fixes #18780

* I added a regression test that reproduces when
using our testsuite with NumPy `1.25.0` with SciPy
`main` (the warning gets turned into an error by
our development test suite)

* I haven't thought too much about the decision
to support this more explicitly, but it sounds
like we've been supporting it for a while and Stefan
seems pretty strongly in favor

* I'm pretty sure this will basically no-op for inputs
that are the expected array type (no extra copy)

+++ scipy/ndimage/_measurements.py
@@ -298,7 +298,7 @@ def find_objects(input, max_label=0):
            [0, 0, 1]])
 
     """
-    input = numpy.asarray(input)
+    input = numpy.asarray(input, dtype=int)
     if numpy.iscomplexobj(input):
         raise TypeError('Complex type not supported')
